requirements.txt
Flask
moviepy
azure-cognitiveservices-speech
ffmpeg-python
